Address PR #407 review: monotonic elapsed clock + generic TX-drop toast
Copilot review findings, both valid:
- writeElapsedMs was measured with System.currentTimeMillis(); a wall
  clock step (NTP set, user change) could fake a fast failure and
  wrongly allow the restart-from-zero UsbRequest fallback after audio
  already went to air. Use SystemClock.elapsedRealtime().
- The tx_audio_dropped toast hard-coded "USB audio device dropped" as
  the cause, but writeAudio can fail for other reasons (libusb setup
  errors, UsbRequest path failures). Genericized to "USB audio error"
  in all 16 locales; the precise cause stays in debug.log via
  describeLibusbWriteError.

Also move the shouldWarnTxDropped tests into their own TxDropWarningTest
file so this PR no longer appends to the same FT8TransmitSignalTest
region as PR #406 (eliminates the guaranteed merge conflict between the
two open PRs).
