+++ ft8af/app/src/main/res/values-fr/strings_compose.xml
@@ -512,5 +512,5 @@
     <string name="settings_hunt_cq_desc">Call CQ when Hunt is idle (no stations heard). When a CQ is detected, automatically stop CQ and answer it.</string>
     <string name="tx_volume_decrease">Decrease TX volume</string>
     <string name="tx_volume_increase">Increase TX volume</string>
-    <string name="tx_audio_dropped">Échec audio TX — aucun signal émis ce cycle (périphérique audio USB déconnecté)</string>
+    <string name="tx_audio_dropped">Échec audio TX — aucun signal émis ce cycle (erreur audio USB)</string>
 </resources>
