+++ ft8af/app/src/main/res/values-it/strings_compose.xml
@@ -544,5 +544,5 @@
     <string name="settings_hunt_cq_desc">Call CQ when Hunt is idle (no stations heard). When a CQ is detected, automatically stop CQ and answer it.</string>
     <string name="tx_volume_decrease">Decrease TX volume</string>
     <string name="tx_volume_increase">Increase TX volume</string>
-    <string name="tx_audio_dropped">Audio TX non riuscito — nessun segnale trasmesso in questo ciclo (dispositivo audio USB disconnesso)</string>
+    <string name="tx_audio_dropped">Audio TX non riuscito — nessun segnale trasmesso in questo ciclo (errore audio USB)</string>
 </resources>
