+++ ft8af/app/src/main/res/values-nl/strings_compose.xml
@@ -555,5 +555,5 @@
     <string name="settings_hunt_cq_desc">Call CQ when Hunt is idle (no stations heard). When a CQ is detected, automatically stop CQ and answer it.</string>
     <string name="tx_volume_decrease">Decrease TX volume</string>
     <string name="tx_volume_increase">Increase TX volume</string>
-    <string name="tx_audio_dropped">TX-audio mislukt — geen signaal verzonden deze cyclus (USB-audioapparaat losgekoppeld)</string>
+    <string name="tx_audio_dropped">TX-audio mislukt — geen signaal verzonden deze cyclus (USB-audiofout)</string>
 </resources>
