+++ ft8af/app/src/main/res/values-ru/strings_compose.xml
@@ -512,5 +512,5 @@
     <string name="settings_hunt_cq_desc">Call CQ when Hunt is idle (no stations heard). When a CQ is detected, automatically stop CQ and answer it.</string>
     <string name="tx_volume_decrease">Decrease TX volume</string>
     <string name="tx_volume_increase">Increase TX volume</string>
-    <string name="tx_audio_dropped">Сбой TX-аудио — в этом цикле сигнал не передан (USB-аудиоустройство отключилось)</string>
+    <string name="tx_audio_dropped">Сбой TX-аудио — в этом цикле сигнал не передан (ошибка USB-аудио)</string>
 </resources>
