+++ ft8af/app/src/main/res/values-zh-rCN/strings_compose.xml
@@ -512,5 +512,5 @@
     <string name="settings_hunt_cq_desc">Call CQ when Hunt is idle (no stations heard). When a CQ is detected, automatically stop CQ and answer it.</string>
     <string name="tx_volume_decrease">Decrease TX volume</string>
     <string name="tx_volume_increase">Increase TX volume</string>
-    <string name="tx_audio_dropped">发射音频失败 — 本周期未发出信号（USB 音频设备断开）</string>
+    <string name="tx_audio_dropped">发射音频失败 — 本周期未发出信号（USB 音频错误）</string>
 </resources>
