+++ ft8af/app/src/main/res/values-zh-rTW/strings_compose.xml
@@ -512,5 +512,5 @@
     <string name="settings_hunt_cq_desc">Call CQ when Hunt is idle (no stations heard). When a CQ is detected, automatically stop CQ and answer it.</string>
     <string name="tx_volume_decrease">Decrease TX volume</string>
     <string name="tx_volume_increase">Increase TX volume</string>
-    <string name="tx_audio_dropped">發射音訊失敗 — 本週期未發出訊號（USB 音訊裝置斷開）</string>
+    <string name="tx_audio_dropped">發射音訊失敗 — 本週期未發出訊號（USB 音訊錯誤）</string>
 </resources>
