+++ ft8af/app/src/main/res/values/strings_compose.xml
@@ -651,5 +651,5 @@
     <string name="language_name_uk" translatable="false">Українська</string>
     <string name="language_name_ar" translatable="false">العربية</string>
 
-    <string name="tx_audio_dropped">TX audio failed — no signal was sent this cycle (USB audio device dropped)</string>
+    <string name="tx_audio_dropped">TX audio failed — no signal was sent this cycle (USB audio error)</string>
 </resources>
